Include audio file restrictions in DEPLOYMENT.rst
While this information is hinted at elsewhere in the documentation, is is only found in pages concerning training models. People who are looking to deploy pre-trained models are likely to miss this information and it should be explicitly stated in parts of the documentation that they are likely to look.